+++ project/taehwan-onboarding/src/main/kotlin/org/survey/taehwanonboarding/domain/entity/response/SurveyResponseItem.kt
@@ -1,4 +1,4 @@
-package org.survey.taehwanonboarding.domain.response
+package org.survey.taehwanonboarding.domain.entity.response
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
@@ -9,7 +9,7 @@ import jakarta.persistence.Id
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
-import org.survey.taehwanonboarding.domain.survey.SurveyItem
+import org.survey.taehwanonboarding.domain.entity.survey.SurveyItem
 
 @Entity
 @Table(name = "response_item")
